fraud-detection-backend/requirements.txt
# Fraud Detection Backend Requirements
# Install: pip install -r requirements.txt

# PyTorch (CPU) - For GPU visit: https://pytorch.org
torch>=2.0.0
torchvision>=0.15.0

# YOLO Object Detection
ultralytics>=8.0.0

# Web Server
flask>=3.0.0
flask-cors>=4.0.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: Production server
# gunicorn>=21.0.0
